Revert: [erp] FiscalVMK: подготовка к деноминации; [server] Подгрузка на клиент библиотеки vmk, если таковая требуется.

+++ erp-logics/src/main/java/lsfusion/erp/region/by/machinery/cashregister/fiscalvmk/ReceiptItem.java
@@ -3,18 +3,19 @@ package lsfusion.erp.region.by.machinery.cashregister.fiscalvmk;
 import java.io.Serializable;
 
 public class ReceiptItem implements Serializable {
+    //todo: необходимо переделать для деноминации цены в double
     public boolean isGiftCard;
-    public double price;
+    public long price;
     public double quantity;
     public String barcode;
     public String name;
-    public double sumPos;
-    public double articleDiscSum;
-    public double bonusSum;
-    public double bonusPaid;
+    public long sumPos;
+    public long articleDiscSum;
+    public long bonusSum;
+    public long bonusPaid;
 
-    public ReceiptItem(boolean isGiftCard, double price, double quantity, String barcode, String name, double sumPos,
-                       double articleDiscSum, double bonusSum, double bonusPaid) {
+    public ReceiptItem(boolean isGiftCard, long price, double quantity, String barcode, String name, long sumPos,
+                       long articleDiscSum, long bonusSum, long bonusPaid) {
         this.isGiftCard = isGiftCard;
         this.price = price;
         this.quantity = quantity;
